Fix applying startup setting when starting in auto mode
simple-pid only applies the `set_auto_mode()` `last_output=` parameter
when going from manual mode to auto mode, and it starts up in auto mode
by default.

In case we also want to start in auto mode, we try to set the wanted
starting output via `set_auto_mode()` but that is ineffective.

Fix that by always starting the simple-pid in manual mode, and switch it
to auto mode during the first loop which causes the wanted startup value
to be actually set.