Remove now unneeded handle.init_phase
The handle.init_phase variable is only used to force calling
set_auto_mode() on the first loop iteration when we are in auto mode.

However, now the simple-pid is started in manual mode always so the
second condition of the if expression is True in that case.

Drop the handle.init_phase variable as unneeded.